Apply DBSCAN on datasets
DBSCAN is not like AP. it needs to change the eps.
eps: specifies how close points should be to each other to be considered a part of a cluster. It means that if the distance between two points is lower or equal to this value (eps), these points are considered neighbors.